API optimize
