--- src/main/java/com/iemr/hwc/repo/brd/BRDIntegrationRepository.java
@@ -1,39 +1,0 @@
-/*
-* AMRIT – Accessible Medical Records via Integrated Technology 
-* Integrated EHR (Electronic Health Records) Solution 
-*
-* Copyright (C) "Piramal Swasthya Management and Research Institute" 
-*
-* This file is part of AMRIT.
-*
-* This program is free software: you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation, either version 3 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License
-* along with this program.  If not, see https://www.gnu.org/licenses/.
-*/
-package com.iemr.hwc.repo.brd;
-
-import java.util.ArrayList;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import com.iemr.hwc.data.uptsu.Referred104Details;
-
-@Repository
-public interface BRDIntegrationRepository extends CrudRepository<Referred104Details, Integer>{
-	
-	@Query(value = "call Pr_104IntegrationBhayaData(:startDate,:endDate)", nativeQuery = true)
-	ArrayList<Object[]> getData(@Param("startDate")String startDate,@Param("endDate")String endDate);
-
-}
